Suppress warnings
* Unused variable.
* Conversion from size_t to int.

+++ src/search.c
@@ -5265,7 +5265,7 @@ search_for_fuzzy_match(
 		    {
 			found_new_match = TRUE;
 			*pos = current_pos;
-			*len = STRLEN(*ptr);
+			*len = (int)STRLEN(*ptr);
 			break;
 		    }
 		}
